Time: 0 ms (100.00%), Space: 7.8 MB (41.57%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int scoreOfString(string s) {
         int ret = 0;
-        for(int i = 0; i < s.size()-1; i++) ret += (abs((int)s[i] - (int) s[i+1]));
+        for(int i = 0; i < s.size()-1; i++) ret += (abs(s[i] - s[i+1]));
         return ret;
     }
 };
